+++ src/Structures/Client.js
@@ -8,7 +8,6 @@ module.exports = class Client extends CommandoClient {
 		super({
 			commandPrefix: process.env.COMMAND_PREFIX,
 			owner: process.env.OWNER_ID,
-			invite: process.env.SUPPORT_SERVER,
 			intents: [Intents.NON_PRIVILEGED, Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_PRESENCES],
 			partials: ['CHANNEL'],
 		});